feat: add string represantion to exc_value
Log record info now for sentry seems is for breadcrumbs.
Eox core Sentry config  is using `before_send` and not `before_breadcrumbs`
so some exception are managed without the log_record.
https://docs.sentry.io/platforms/python/guides/django/configuration/filtering/hints/
So the self.exc_text is ''.
If this is empty regexes dont make anything in
https://github.com/eduNEXT/eox-core/blob/v6.1.0/eox_core/integrations/sentry.py#L154
so this change add `exc_value` if nothing is in log_records, the idea is to could compare
at least with the str of the `exc_value`.