feat: add pkce-openid-backend
This add a generic backend based in ConfigurableOpenIdConnectAuth but
with PKCE.
This backend is inspired  in the social-core way to implement PKCE.
There is a current PR in working, but for the moment, that class is not merged and accesible.
So after that is finished this has it code for `code_challenge` and `code_challenge_method`implementation.
PR: https://github.com/python-social-auth/social-core/pull/856