feat: add pkce-openid-backend
This adds a generic backend based on ConfigurableOpenIdConnectAuth but
with PKCE.
This backend is inspired in the social-core way to implement PKCE.
There is a current PR in work, but for the moment, that class is not merged and accessible.
So after that is finished this has it code for `code_challenge` and
`code_challenge_method`implementation.
PR: https://github.com/python-social-auth/social-core/pull/856

Co-authored-by: Omar Al-Ithawi @ NELC <134635705+omar-nelc@users.noreply.github.com>
Co-authored-by: Andrey Cañon <andrey.canon@edunext.co>